// packages/cfs-lib/src/types/code-generation.ts
/**
 *
 * Copyright (c) 2025 Analog Devices, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

export type CodeGenerationResult = (string | CodeGenerationFailure)[];

export interface CodeGenerationFailure {
	name: string;
	error: string;
}

export interface CodeGenerationProject {
	projectId: string;
	includeAI?: boolean;
}